+++ xaux/automation/__init__.py
@@ -1,0 +1,9 @@
+# copyright ############################### #
+# This file is part of the Xaux Package.    #
+# Copyright (c) CERN, 2025.                 #
+# ######################################### #
+
+from .template import JobTemplate
+from .da import DAJob
+from .collimation import LossMapPencilJob
+from .manager import JobManager
